+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+-   Added `enableCrossOrigin` prop to `Tumbnail` component
+-   Added `crossOrigin` prop to `Tumbnail` component
+-   Added `updateOnWindowResize` prop to `Tumbnail` component
+-   Added `resizeDebounceTime` prop to `Tumbnail` component
+
 ## [0.21.13][] - 2020-03-23
 
 ### Fixed